requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
google-api-core==1.30.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
six==1.16.0
threaded==4.1.0
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.5
python-dotenv==0.19.0
PureCloudPlatformClientV2==132.0.0